+++ Backend/apps/informacion/Views/TutoriaConcluidaView.py
@@ -126,22 +126,23 @@ class TutoriaConcluidaViewSet(viewsets.ModelViewSet):
             
             #eliminar imagen en el bucket de clouflare
             # extraer solo el nombre de la imagen
-            image_path = tutoria_concluida.image_r2.split("/")[-1]
-
-            try:
-                s3.delete_object(Bucket=settings.R2_BUCKET_NAME, Key=image_path)
-            except Exception as e:
-                return Response({"error": f"No se pudo eliminar la imagen en R2: {str(e)}"},
-                            status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            if tutoria_concluida.image_r2:
+                image_path = tutoria_concluida.image_r2.split("/")[-1]
+                try:
+                    s3.delete_object(Bucket=settings.R2_BUCKET_NAME, Key=image_path)
+                except Exception as e:
+                    return Response({"error": f"No se pudo eliminar la imagen en R2: {str(e)}"},
+                                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            
             #eliminar archivo en el bucket de clouflare
             # extraer solo el nombre del archivo
-            file_path = tutoria_concluida.file_r2.split("/")[-1]
-
-            try:
-                s3.delete_object(Bucket=settings.R2_BUCKET_FILES_NAME, Key=file_path)
-            except Exception as e:
-                return Response({"error": f"No se pudo eliminar el archivo en R2: {str(e)}"},
-                                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            if tutoria_concluida.file_r2:
+                file_path = tutoria_concluida.file_r2.split("/")[-1]
+                try:
+                    s3.delete_object(Bucket=settings.R2_BUCKET_FILES_NAME, Key=file_path)
+                except Exception as e:
+                    return Response({"error": f"No se pudo eliminar el archivo en R2: {str(e)}"},
+                                    status=status.HTTP_500_INTERNAL_SERVER_ERROR)
             
             tutoria_concluida.delete()
             return Response({'Tutoría concluida eliminada correctamente'}, status=status.HTTP_204_NO_CONTENT)
